Refactor RenderViewModel into specific View
